+++ docs/CHANGELOG.md
@@ -42,6 +42,8 @@ Version 1.9.0 (not released yet)
 - Add support for bluebeard.bty (sound config file) in mods
 - Properly handle WM_PAINT in dedicated server, may improve performance (DF bug)
 - Disable adding faces to fix PS2 tiling in editor
+- Disable red background in log window if geometry limits are crossed in editor
+- Update geometry limits in editor so they show static pools size (crossing it can have tiny performance impact)
 
 [@GooberRF](https://github.com/GooberRF)
 - Fix crash when `verify_level` command is run without a level being loaded